fix: 批量修复 pandas 3.0 / numpy 2.0 / scipy 1.15 / xarray 2025.4 兼容性
1. float16 -> float32(reptile.py, scipy.sparse 不支持 float16)
2. np.NaN -> np.nan(numpy 2.0 移除,4 处)
3. np.product monkey-patch(zarr 2.13.3 用了已移除的 np.product)
4. DMR .to_pandas().astype(str) -> .to_numpy().astype(object)
   (pandas 3.0 astype(str) 返回 StringDtype,xarray/zarr 不接受;
    xarray 2025.4 所有字符串 coord 都是 StrDtype,需转 object dtype)
5. dtype=="O" -> str(dtype) in ("object","str")
   (pandas 3.0 StringDtype 的 str() 是 "str" 不是 "O",5 处)
6. groupby(axis=1) -> .T.groupby()(pandas 3.0 移除 axis 参数)
7. astype(str) -> astype(object)(xarray coord,避免 StrDtype)
8. mpileup 短行 fields 越界 -> len(fields)<7: continue
9. xarray to_numpy(dtype=) -> .to_numpy().astype()
10. bismark version regex 兼容 v3(step4_report_rna_met.py)

兼容环境: pandas>=3.0, numpy>=2.0, scipy>=1.15, xarray>=2025.4
旧环境用户可继续用 v1.2.0 tag。

+++ ALLCools/reptile/reptile.py
@@ -186,7 +186,7 @@ def _predict_sample(
             enhancer_proba.fillna(0, inplace=True)
             total_proba.append(enhancer_proba)
 
-        total_proba = pd.DataFrame({sample: pd.concat(total_proba).astype(np.float16)})
+        total_proba = pd.DataFrame({sample: pd.concat(total_proba).astype(np.float32)})
         # mask small values
         total_proba[total_proba < mask_cutoff] = 0
         total_proba.index.name = region_ds.region_dim
